refactroing. split system wrappers to .h and .cc using unrestricted updates in diagram to do discrete time control
