+++ drake/examples/QPInverseDynamicsForHumanoids/system/qp_controller_system.h
@@ -1,11 +1,9 @@
 #pragma once
 
-#include <iostream>
 #include <memory>
-#include <utility>
 
-#include "drake/examples/QPInverseDynamicsForHumanoids/lcm_utils.h"
 #include "drake/examples/QPInverseDynamicsForHumanoids/qp_controller.h"
+#include "drake/multibody/rigid_body_tree.h"
 #include "drake/systems/framework/leaf_system.h"
 
 namespace drake {
@@ -21,71 +19,36 @@ namespace qp_inverse_dynamics {
  */
 class QPControllerSystem : public systems::LeafSystem<double> {
  public:
-  explicit QPControllerSystem(const RigidBodyTree<double>& robot)
-      : robot_(robot) {
-    input_port_index_humanoid_status_ = DeclareAbstractInputPort().get_index();
-    input_port_index_qp_input_ = DeclareAbstractInputPort().get_index();
-    output_port_index_qp_input_ = DeclareAbstractOutputPort().get_index();
+  explicit QPControllerSystem(const RigidBodyTree<double>& robot);
 
-    DRAKE_ASSERT(this->get_num_input_ports() == 2);
-    DRAKE_ASSERT(this->get_num_output_ports() == 1);
+  void DoCalcOutput(const systems::Context<double>& context,
+                    systems::SystemOutput<double>* output) const override;
 
-    set_name("qp_controller");
-  }
-
-  void DoCalcOutput(const Context<double>& context,
-                    SystemOutput<double>* output) const override {
-    // Inputs:
-    const HumanoidStatus* rs = EvalInputValue<HumanoidStatus>(
-        context, input_port_index_humanoid_status_);
-
-    const lcmt_qp_input* qp_input_msg =
-        EvalInputValue<lcmt_qp_input>(context, input_port_index_qp_input_);
-
-    QpInput qp_input;
-    DecodeQpInput(robot_, *qp_input_msg, &qp_input);
-
-    // Output:
-    QpOutput& qp_output = output->GetMutableData(output_port_index_qp_input_)
-                              ->GetMutableValue<QpOutput>();
-
-    if (qp_controller_.Control(*rs, qp_input, &qp_output) < 0) {
-      std::cout << rs->position().transpose() << std::endl;
-      std::cout << rs->velocity().transpose() << std::endl;
-      std::cout << qp_input << std::endl;
-      throw std::runtime_error("System2QP: QP cannot solve\n");
-    }
-  }
-
-  std::unique_ptr<SystemOutput<double>> AllocateOutput(
-      const Context<double>& context) const override {
-    std::unique_ptr<LeafSystemOutput<double>> output(
-        new LeafSystemOutput<double>);
-    QpOutput out(GetDofNames(robot_));
-    output->add_port(std::unique_ptr<AbstractValue>(new Value<QpOutput>(out)));
-    return std::move(output);
-  }
+  std::unique_ptr<systems::AbstractValue> AllocateOutputAbstract(
+      const systems::OutputPortDescriptor<double>& descriptor) const override;
 
   /**
    * @return Port for the input: HumanoidStatus.
    */
-  inline const InputPortDescriptor<double>& get_input_port_humanoid_status()
-      const {
+  inline const systems::InputPortDescriptor<double>&
+  get_input_port_humanoid_status() const {
     return get_input_port(input_port_index_humanoid_status_);
   }
 
   /**
    * @return Port for the input: QpInput.
    */
-  inline const InputPortDescriptor<double>& get_input_port_qp_input() const {
+  inline const systems::InputPortDescriptor<double>& get_input_port_qp_input()
+      const {
     return get_input_port(input_port_index_qp_input_);
   }
 
   /**
    * @return Port for the output: QpOutput.
    */
-  inline const OutputPortDescriptor<double>& get_output_port_qp_output() const {
-    return get_output_port(output_port_index_qp_input_);
+  inline const systems::OutputPortDescriptor<double>&
+  get_output_port_qp_output() const {
+    return get_output_port(output_port_index_qp_output_);
   }
 
  private:
@@ -100,7 +63,7 @@ class QPControllerSystem : public systems::LeafSystem<double> {
 
   int input_port_index_humanoid_status_;
   int input_port_index_qp_input_;
-  int output_port_index_qp_input_;
+  int output_port_index_qp_output_;
 };
 
 }  // namespace qp_inverse_dynamics
