+++ drake/examples/QPInverseDynamicsForHumanoids/test/CMakeLists.txt
@@ -1,8 +1,9 @@
 drake_add_cc_test(setpoint_test)
-target_link_libraries(setpoint_test drakeQPController)
+target_link_libraries(setpoint_test
+  drakeMath)
 
 if(lcm_FOUND AND yaml-cpp_FOUND)
-  drake_add_cc_test(NAME lcm_utils_test)
+  drake_add_cc_test(lcm_utils_test)
   target_link_libraries(lcm_utils_test
     drakeMultibodyParsers
     drakeQPController
@@ -18,17 +19,3 @@ if(gurobi_FOUND AND yaml-cpp_FOUND)
     drakeRigidBodyAliasGroups
     drakeValkyrieConstants)
 endif()
-
-if(gurobi_FOUND AND lcm_FOUND)
-  add_executable(valkyrie_qp_inverse_dynamics_controller
-    valkyrie_qp_inverse_dynamics_controller.cc)
-  target_link_libraries(valkyrie_qp_inverse_dynamics_controller
-    drakeLCMSystem2
-    drakeMultibodyParsers
-    drakeQPController
-    drakeQPControllerLcmTranslator
-    drakeQPControllerParamParsing
-    drakeRigidBodyAliasGroups
-    drakeValkyrieConstants)
-endif()
-
